Gut and reorganise all lib files
