--- lib/Checks.groovy
@@ -1,66 +1,0 @@
-/*
- * This file holds several functions used to perform standard checks for the nf-core pipeline template.
- */
-
-import org.yaml.snakeyaml.Yaml
-
-class Checks {
-
-    public static void checkCondaChannels(log) {
-        Yaml parser = new Yaml()
-        def channels = []
-        try {
-            def config = parser.load("conda config --show channels".execute().text)
-            channels = config.channels
-        } catch(NullPointerException | IOException e) {
-            log.warn "Could not verify conda channel configuration."
-            return
-        }
-
-        // Check that all channels are present
-        def required_channels = ['conda-forge', 'bioconda', 'defaults']
-        def conda_check_failed = !required_channels.every { ch -> ch in channels }
-
-        // Check that they are in the right order
-        conda_check_failed |= !(channels.indexOf('conda-forge') < channels.indexOf('bioconda'))
-        conda_check_failed |= !(channels.indexOf('bioconda') < channels.indexOf('defaults'))
-
-        if (conda_check_failed) {
-            log.warn "=============================================================================\n" +
-                     "  There is a problem with your Conda configuration!\n\n" + 
-                     "  You will need to set-up the conda-forge and bioconda channels correctly.\n" +
-                     "  Please refer to https://bioconda.github.io/user/install.html#set-up-channels\n" +
-                     "  NB: The order of the channels matters!\n" +
-                     "==================================================================================="
-        }
-    }
-
-    public static void awsBatch(workflow, params) {
-        if (workflow.profile.contains('awsbatch')) {
-            assert (params.awsqueue && params.awsregion) : "Specify correct --awsqueue and --awsregion parameters on AWSBatch!"
-            // Check outdir paths to be S3 buckets if running on AWSBatch
-            // related: https://github.com/nextflow-io/nextflow/issues/813
-            assert params.outdir.startsWith('s3:')       : "Outdir not on S3 - specify S3 Bucket to run on AWSBatch!"
-            // Prevent trace files to be stored on S3 since S3 does not support rolling files.
-            assert !params.tracedir.startsWith('s3:')    :  "Specify a local tracedir or run without trace! S3 cannot be used for tracefiles."
-        }
-    }
-
-    public static void hostName(workflow, params, log) {
-        Map colors = Utils.logColours(params.monochrome_logs)
-        if (params.hostnames) {
-            def hostname = "hostname".execute().text.trim()
-            params.hostnames.each { prof, hnames ->
-                hnames.each { hname ->
-                    if (hostname.contains(hname) && !workflow.profile.contains(prof)) {
-                        log.info "=${colors.yellow}====================================================${colors.reset}=\n" +
-                                  "${colors.yellow}WARN: You are running with `-profile $workflow.profile`\n" +
-                                  "      but your machine hostname is ${colors.white}'$hostname'${colors.reset}.\n" +
-                                  "      ${colors.yellow_bold}Please use `-profile $prof${colors.reset}`\n" +
-                                  "=${colors.yellow}====================================================${colors.reset}="
-                    }
-                }
-            }
-        }
-    }
-}
