+++ lib/WorkflowMain.groovy
@@ -1,0 +1,74 @@
+/*
+ *This file holds several functions specific to the main.nf workflow in the nf-core/rnaseq pipeline
+ */
+
+class WorkflowMain {
+
+    // Citation string
+    public static String citation(workflow) {
+        return "If you use ${workflow.manifest.name} for your analysis please cite:\n\n" +
+               "* The pipeline\n" + 
+               "  https://doi.org/10.5281/zenodo.1400710\n\n" +
+               "* The nf-core framework\n" +
+               "  https://doi.org/10.1038/s41587-020-0439-x\n\n" +
+               "* Software dependencies\n" +
+               "  https://github.com/${workflow.manifest.name}/blob/master/CITATIONS.md"
+    }
+
+    // Print help to screen if required
+    public static String help(workflow, params, log) {
+        def command = "nextflow run nf-core/rnaseq --input samplesheet.csv --genome GRCh37 -profile docker"
+        def help_string = ''
+        help_string += NfcoreTemplate.logo(workflow, params.monochrome_logs)
+        help_string += NfcoreSchema.paramsHelp(workflow, params, command)
+        help_string += '\n' + citation(workflow) + '\n'
+        help_string += NfcoreTemplate.dashedLine(params.monochrome_logs)
+        return help_string
+    }
+
+    // Print parameter summary log to screen
+    public static String paramsSummaryLog(workflow, params, log) {
+        def summary_log = ''
+        summary_log += NfcoreTemplate.logo(workflow, params.monochrome_logs)
+        summary_log += NfcoreSchema.paramsSummaryLog(workflow, params)
+        summary_log += '\n' + citation(workflow) + '\n'
+        summary_log += NfcoreTemplate.dashedLine(params.monochrome_logs)
+        return summary_log
+    }
+
+    // Validate parameters and print summary to screen
+    public static void initialise(workflow, params, log) {
+        // Print help to screen if required
+        if (params.help) {
+            log.info help(workflow, params, log)
+            System.exit(0)
+        }
+
+        // Validate workflow parameters via the JSON schema
+        if (params.validate_params) {
+            NfcoreSchema.validateParameters(workflow, params, log)
+        }
+
+        // Print parameter summary log to screen
+        log.info paramsSummaryLog(workflow, params, log)
+
+        // Check that conda channels are set-up correctly
+        if (params.enable_conda) {
+            Utils.checkCondaChannels(log)
+        }
+
+        // Check AWS batch settings
+        NfcoreTemplate.awsBatch(workflow, params)
+
+        // Check the hostnames against configured profiles
+        NfcoreTemplate.hostName(workflow, params, log)
+
+        // Check at least one form of input has been provided
+        if (!params.public_data_ids && !params.input) {
+            log.error "Please specify at least one form of input for the pipeline e.g. '--input samplsheet.csv' or '--public_data_ids ids.txt'."
+            System.exit(1)
+        }
+    }
+}
+
+
